requirements.txt
django>=1.4,<1.5
django-crispy-forms>=1.2,<1.3
South==0.7.6
django-autoslug==1.4.1
